feat(pre-processing): Unify Spec Tags Before Generation
Unify spec tags to the spec info-title before generating the sdk
