+++ openapi/src/main/kotlin/com/expediagroup/sdk/generators/openapi/Constant.kt
@@ -13,7 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.expediagroup.sdk.generators.yaml.processor
+package com.expediagroup.sdk.generators.openapi
 
-class PreProcessingException(override val message: String, override val cause: Throwable? = null) :
-    RuntimeException(message, cause)
+internal object Constant {
+    const val INDENTATION_LENGTH = 2
+    const val INDICATOR_INDENTATION_LENGTH = 2
+    const val INDENT_WITH_INDICATOR = true
+    val NON_ALPHANUMERIC_REGEX = Regex("[^a-z0-9]")
+}
